+++ src/custom.d.ts
@@ -2,10 +2,12 @@ declare module '*.css';
 declare module '*.scss';
 declare module '*.sass';
 declare module '@d3fc/d3fc-sample';
+
 declare module '*.graphql' {
     const value: string;
     export = value;
 }
+
 declare module "codegen.*.json" {
     const value: {
         schema: string,